lib update

+++ libplateau/CSharpPLATEAU/Geometries/GeoReference.cs
@@ -142,7 +142,7 @@ namespace PLATEAU.Geometries
             return outLatLon;
         }
 
-        public PlateauVector3d Convert(PlateauVector3d point, bool convertAxis, double epsg)
+        public PlateauVector3d Convert(PlateauVector3d point, bool convertAxis, int epsg)
         {
             var result = NativeMethods.plateau_geo_reference_convert(
                 Handle, out var outXyz,
@@ -256,7 +256,7 @@ namespace PLATEAU.Geometries
             internal static extern APIResult plateau_geo_reference_convert(
             [In] IntPtr geoReferencePtr,
             out PlateauVector3d outXyz,
-            PlateauVector3d point, [MarshalAs(UnmanagedType.U1)] bool convertAxis, double epsg);
+            PlateauVector3d point, [MarshalAs(UnmanagedType.U1)] bool convertAxis, int epsg);
 
             [DllImport(DLLUtil.DllName)]
             internal static extern APIResult plateau_geo_reference_get_reference_point(
